+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
 ### Added
+- Matlab Mesh object has now a `transform` method to apply affine
+  transformations to the mesh
 - Option to ignore samples outside the scan `ignoreExteriorSamples`, see [#52](https://github.com/ithron/CortidQCT/issues/52))
 - CI builds for clang-7 and macOS (xcode 10.2)
 - Ray-mesh intersection computation